Remove outdated (and broken) prolog dep step. (#188)
The pack_install(terminus_store_prolog) step appears to be no longer necessary. (It also fails on nixos, and likely any other system that needs to pass env vars to cargo). 

The pack_install(tus) step can be replaced with the equivalent makefile target. Meaning there is no need to manually run swipl at all.

The instructions were also missing the target to install the dashboard, which I found to be a point of confusion.